fix bazel
